+++ include/SCC/lexical_analyzer.h
@@ -4,7 +4,7 @@
 #include <filesystem>
 #include <sstream>
 
-#include "SCC/ast.h"
+#include "SCC/ast/ast.h"
 #include "SCC/config/scc_config.h"
 #include "SCC/log.h"
