+++ include/SCC/query_assembler.h
@@ -7,7 +7,7 @@
 #include <tuple>
 #include <sstream>
 
-#include "SCC/ast.h"
+#include "SCC/ast/ast.h"
 #include "SCC/config/scc_config.h"
 #include "SCC/log.h"
 #include "SCC/syntax_analyzer.h"
