+++ test/SQL_AST_test/SQL_AST_test.cpp
@@ -1,6 +1,6 @@
 #include "gtest/gtest.h"
 #include "gmock/gmock.h"
-#include "SCC/ast.h"
+#include "SCC/ast/ast.h"
 #include "SCC/lexical_analyzer.h"
 #include "SCC/syntax_analyzer.h"
 #include "SCC/tree_dump.h"
